# api/openapi.yaml
openapi: 3.1.0
info:
  title: Quantum Control API
  version: 0.3.0-alpha.1
  description: Public typed administration API with identity, immutable planning, human confirmation, transactional service mutation and durable audit contracts.
servers:
  - url: http://127.0.0.1:17440
paths:
  /healthz:
    get:
      summary: Public process liveness
      responses:
        '200': {description: Process is alive}
  /readyz:
    get:
      summary: Broker and security readiness
      responses:
        '200': {description: Broker and security subsystem are ready}
        '503':
          description: Broker or security subsystem unavailable
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ErrorEnvelope'}
  /v1/control/info:
    get:
      summary: Product and capability information
      security: [{bearerAuth: []}]
      responses:
        '200': {description: Control information}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
  /v1/components:
    get:
      summary: Read component inventory snapshot
      description: Requires inventory.read. Fixed read-only probes only.
      security: [{bearerAuth: []}]
      responses:
        '200':
          description: Component inventory snapshot
          content:
            application/json:
              schema: {$ref: '#/components/schemas/InventorySnapshot'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '500':
          description: Inventory could not be completed safely
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ErrorEnvelope'}
  /v1/components/{id}:
    get:
      summary: Read one canonical component inventory record
      security: [{bearerAuth: []}]
      parameters:
        - in: path
          name: id
          required: true
          schema: {type: string, pattern: '^[a-z0-9][a-z0-9-]{0,63}$'}
      responses:
        '200':
          description: Component inventory record
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ComponentInventory'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404':
          description: Component ID is not part of the inventory contract
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ErrorEnvelope'}
  /v1/operations:
    get:
      summary: Allowlisted operation catalog
      security: [{bearerAuth: []}]
      responses:
        '200': {description: Operation definitions}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '502': {$ref: '#/components/responses/BrokerError'}
  /v1/operations/plan:
    post:
      summary: Create an immutable expiring operation plan or TCI proposal
      description: Requires operations.plan. Caller actor fields are overwritten by authenticated identity.
      security: [{bearerAuth: []}]
      parameters:
        - in: header
          name: X-Quantum-Session-ID
          required: false
          schema: {type: string, maxLength: 160}
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/OperationRequest'}
      responses:
        '200':
          description: Valid immutable plan
          content:
            application/json:
              schema: {$ref: '#/components/schemas/OperationPlan'}
        '400':
          description: Invalid request or rejected plan
          content:
            application/json:
              schema: {$ref: '#/components/schemas/OperationPlan'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '413': {description: Request body exceeds the configured limit}
        '502': {$ref: '#/components/responses/BrokerError'}
  /v1/operations/execute:
    post:
      summary: Execute a read-only allowlisted operation
      description: Requires operations.execute.readonly. Confirmation-required mutations are rejected on this route.
      security: [{bearerAuth: []}]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/OperationRequest'}
      responses:
        '200': {description: Read-only operation completed}
        '400': {description: Operation rejected by policy or validation}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '413': {description: Request body exceeds the configured limit}
        '500': {description: Allowlisted read operation failed}
        '502': {$ref: '#/components/responses/BrokerError'}
  /v1/confirmations:
    post:
      summary: Issue a short-lived single-use confirmation grant
      description: Requires operations.confirm and an authenticated human approver. qcored independently authenticates the approver and owns grant state.
      security: [{bearerAuth: []}]
      parameters:
        - in: header
          name: X-Quantum-Session-ID
          required: false
          schema: {type: string, maxLength: 160}
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [plan_id]
              properties:
                plan_id: {type: string, maxLength: 160}
      responses:
        '201':
          description: Confirmation grant issued. Raw token is returned once and only its SHA-256 digest is persisted.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/GrantResponse'}
        '400': {description: Plan is not confirmable under current broker policy}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {description: Plan missing or expired}
        '503': {description: Broker confirmation subsystem unavailable}
  /v1/operations/execute-approved:
    post:
      summary: Execute one approved confirmation-required service mutation
      description: Requires operations.execute.mutate. qcored independently authenticates the executor, revalidates the exact plan and consumes the grant before invoking the privileged adapter.
      security: [{bearerAuth: []}]
      parameters:
        - in: header
          name: X-Quantum-Session-ID
          required: false
          schema: {type: string, maxLength: 160}
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/ApprovedExecutionRequest'}
      responses:
        '200':
          description: Approved service mutation completed and postcondition verified
          content:
            application/json:
              schema: {$ref: '#/components/schemas/OperationResponse'}
        '400':
          description: Plan/grant rejected or operation semantically rejected
          content:
            application/json:
              schema: {$ref: '#/components/schemas/OperationResponse'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {description: Plan missing or expired}
        '500':
          description: Privileged action or required postcondition failed
          content:
            application/json:
              schema: {$ref: '#/components/schemas/OperationResponse'}
        '502': {$ref: '#/components/responses/BrokerError'}
        '503': {description: Approved mutation subsystem unavailable}
  /v1/audit:
    get:
      summary: Read permission-scoped durable audit history
      description: Requires audit.read. No audit mutation method exists.
      security: [{bearerAuth: []}]
      parameters:
        - in: query
          name: limit
          schema: {type: integer, minimum: 1, maximum: 500, default: 100}
        - in: query
          name: actor_id
          schema: {type: string, maxLength: 128}
        - in: query
          name: action
          schema: {type: string, maxLength: 160}
      responses:
        '200': {description: Audit records and chain integrity metadata}
        '400': {description: Invalid audit filter}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '503': {description: Durable audit unavailable}
  /v1/audit/integrity:
    get:
      summary: Read durable audit chain head and verification state
      security: [{bearerAuth: []}]
      responses:
        '200': {description: Audit integrity metadata}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '503': {description: Durable audit unavailable}
  /v1/system/status:
    get:
      summary: Read bounded system status
      security: [{bearerAuth: []}]
      responses:
        '200': {description: System snapshot}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '502': {$ref: '#/components/responses/BrokerError'}
  /v1/services/{unit}:
    get:
      summary: Read systemd unit state
      security: [{bearerAuth: []}]
      parameters:
        - in: path
          name: unit
          required: true
          schema: {type: string, pattern: '^[A-Za-z0-9][A-Za-z0-9_.@:-]{0,127}$'}
      responses:
        '200': {description: Unit state}
        '400': {description: Unit identifier rejected}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '502': {$ref: '#/components/responses/BrokerError'}
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: Actor token. Remote listeners require a credential source. Loopback may use a restricted read-only bootstrap actor when no credential source is configured.
  responses:
    Unauthorized:
      description: Valid actor credential required
      content:
        application/json:
          schema: {$ref: '#/components/schemas/ErrorEnvelope'}
    Forbidden:
      description: Authenticated actor lacks the required permission
      content:
        application/json:
          schema: {$ref: '#/components/schemas/ErrorEnvelope'}
    BrokerError:
      description: Privileged broker transport failed or produced an ambiguous result
      content:
        application/json:
          schema: {$ref: '#/components/schemas/ErrorEnvelope'}
  schemas:
    Actor:
      type: object
      required: [id, kind, display_name, roles, permissions]
      properties:
        id: {type: string}
        kind: {type: string, enum: [human, service, tci]}
        display_name: {type: string}
        roles: {type: array, items: {type: string}}
        permissions: {type: array, items: {type: string}}
    OperationRequest:
      type: object
      required: [action]
      additionalProperties: false
      properties:
        request_id: {type: string}
        action: {type: string}
        actor:
          type: string
          description: Ignored and replaced by authenticated actor identity.
        parameters:
          type: object
          additionalProperties: {type: string}
        confirmation:
          type: string
          description: Legacy field. Cleared by the public service and never sufficient for confirmation-required execution.
    OperationPlan:
      type: object
      required: [schema, id, digest, request_id, session_id, actor, action, parameters, risk, requires_confirmation, valid, created_at, expires_at]
      properties:
        schema: {type: string, const: quantum.control/operation-plan/v1alpha1}
        id: {type: string}
        digest: {type: string, pattern: '^[0-9a-f]{64}$'}
        request_id: {type: string}
        session_id: {type: string}
        actor: {$ref: '#/components/schemas/Actor'}
        action: {type: string}
        parameters:
          type: array
          items:
            type: object
            required: [name, value]
            properties:
              name: {type: string}
              value: {type: string}
        risk: {type: string, enum: [read-only, low, high, destructive]}
        requires_confirmation: {type: boolean}
        valid: {type: boolean}
        created_at: {type: string, format: date-time}
        expires_at: {type: string, format: date-time}
        error_code: {type: string}
    ConfirmationGrant:
      type: object
      required: [schema, id, plan_id, plan_digest, subject_actor_id, session_id, approver, action, issued_at, expires_at]
      properties:
        schema: {type: string, const: quantum.control/confirmation-grant/v1alpha1}
        id: {type: string}
        plan_id: {type: string}
        plan_digest: {type: string, pattern: '^[0-9a-f]{64}$'}
        subject_actor_id: {type: string}
        session_id: {type: string}
        approver: {$ref: '#/components/schemas/Actor'}
        action: {type: string}
        issued_at: {type: string, format: date-time}
        expires_at: {type: string, format: date-time}
        used_at: {type: string, format: date-time}
    GrantResponse:
      type: object
      required: [grant, token]
      properties:
        grant: {$ref: '#/components/schemas/ConfirmationGrant'}
        token:
          type: string
          description: One-time raw token returned to the caller. Only its SHA-256 digest is persisted by qcored.
    ApprovedExecutionRequest:
      type: object
      additionalProperties: false
      required: [plan_id, confirmation_token]
      properties:
        plan_id: {type: string, maxLength: 160}
        confirmation_token: {type: string, maxLength: 512}
    OperationResponse:
      type: object
      required: [request_id, action, status, risk, started_at, finished_at, audit_id]
      properties:
        request_id: {type: string}
        action: {type: string}
        status: {type: string, enum: [completed, rejected, failed]}
        risk: {type: string, enum: [read-only, low, high, destructive]}
        started_at: {type: string, format: date-time}
        finished_at: {type: string, format: date-time}
        audit_id: {type: string}
        result:
          type: object
          additionalProperties: true
          description: Bounded operation result including service pre/postcondition and recovery_status when applicable.
        error: {$ref: '#/components/schemas/Problem'}
    InventorySnapshot:
      type: object
      required: [schema, observed_at, components]
      additionalProperties: false
      properties:
        schema: {type: string, const: quantum.control/component-inventory/v1alpha1}
        observed_at: {type: string, format: date-time}
        components:
          type: array
          items: {$ref: '#/components/schemas/ComponentInventory'}
    ComponentInventory:
      type: object
      required: [id, name, category, version, ownership, health, services, listeners, roots, capabilities, evidence, observed_at, warnings]
      additionalProperties: false
      properties:
        id: {type: string, pattern: '^[a-z0-9][a-z0-9-]{0,63}$'}
        name: {type: string}
        category: {type: string}
        version: {type: string}
        ownership: {type: string, enum: [managed, external, disabled, unknown]}
        health: {type: string, enum: [healthy, degraded, disabled, unknown]}
        services: {type: array, items: {type: string}}
        listeners:
          type: array
          items:
            type: object
            required: [address, port, protocol]
            properties:
              address: {type: string}
              port: {type: integer, minimum: 1, maximum: 65535}
              protocol: {type: string}
        roots: {type: array, items: {type: string}}
        capabilities: {type: array, items: {type: string}}
        evidence:
          type: array
          items:
            type: object
            required: [kind, source]
            properties:
              kind: {type: string}
              source: {type: string}
              detail: {type: string}
        observed_at: {type: string, format: date-time}
        warnings: {type: array, items: {type: string}}
    Problem:
      type: object
      required: [code, message]
      properties:
        code: {type: string}
        message: {type: string}
    ErrorEnvelope:
      type: object
      required: [error, request_id]
      properties:
        error: {$ref: '#/components/schemas/Problem'}
        request_id: {type: string}